--- src/HelmerDemo.BlazorServer.Application/Events/TimeTimer.cs
@@ -1,16 +1,0 @@
-﻿using HelmerDemo.BlazorServer.Application.Domain;
-
-namespace HelmerDemo.BlazorServer.Application.Events
-{
-    public class TimeTimer : ITimeTimer
-	{
-		/// <inheritdoc />
-		public event EventHandler<TimeTimerEventArgs> OnTimeChanged = delegate { };
-
-		/// <inheritdoc />
-		public void RaiseTimeUpdate(DigitalClock time)
-		{
-			if (OnTimeChanged != null) OnTimeChanged(this, new TimeTimerEventArgs(time));
-		}
-	}
-}
